refactor: move `mongoObjectId` scalar & `paginationOptions` dto to `global/dto` folder
