+++ src/app.module.ts
@@ -5,7 +5,7 @@ import { PostsModule } from './posts/posts.module';
 import { join } from 'path';
 import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
 import { GraphQLModule } from '@nestjs/graphql';
-import { MongoObjectIdScalar } from './global-dto/mongoObjectId.scalar';
+import { MongoObjectIdScalar } from './global/dto/mongoObjectId.scalar';
 import { AuthModule } from './auth/auth.module';
 
 @Module({
