+++ src/posts/dto/update-post.input.ts
@@ -1,6 +1,6 @@
 import { InputType, Field } from '@nestjs/graphql';
 import { Types } from 'mongoose';
-import { MongoObjectIdScalar } from '../../global-dto/mongoObjectId.scalar';
+import { MongoObjectIdScalar } from '../../global/dto/mongoObjectId.scalar';
 import { PostThumbnailInterface } from '../interfaces';
 
 @InputType()
