+++ src/posts/posts.resolver.ts
@@ -13,9 +13,9 @@ import { EditPostInput } from './dto/update-post.input';
 import { SearchForPostInput } from './dto/search-for-post.input';
 import { User } from '../global-dto/user-data.type';
 import { UsersService } from '../users/users.service';
-import { PaginationOptions } from '../global-dto/pagination-options.dto';
+import { PaginationOptions } from '../global/dto/pagination-options.dto';
 import { Types } from 'mongoose';
-import { MongoObjectIdScalar } from '../global-dto/mongoObjectId.scalar';
+import { MongoObjectIdScalar } from '../global/dto/mongoObjectId.scalar';
 import { UseGuards } from '@nestjs/common';
 import { CurrentUser, GqlJwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 import { AuthorizedUser } from '../auth/dto/auth-related.dto';
