+++ src/users/users.resolver.ts
@@ -13,8 +13,8 @@ import { SearchForUserInput } from './dto/search-for-users.input';
 import { User } from '../global-dto/user-data.type';
 import { Post } from '../global-dto/post-data.type';
 import { UserDocument } from './schemas/user.schema';
-import { PaginationOptions } from '../global-dto/pagination-options.dto';
-import { MongoObjectIdScalar } from '../global-dto/mongoObjectId.scalar';
+import { PaginationOptions } from '../global/dto/pagination-options.dto';
+import { MongoObjectIdScalar } from '../global/dto/mongoObjectId.scalar';
 import { Types } from 'mongoose';
 import { UseGuards } from '@nestjs/common';
 import { CurrentUser, GqlJwtAuthGuard } from '../auth/guards/jwt-auth.guard';
